standardize safe_import as a util
