+++ python/lancedb/utils/general.py
@@ -443,22 +443,3 @@ def threaded_request(
         ).start()
     else:
         return func(*args, **kwargs)
-
-
-def safe_import(module: str, mitigation=None):
-    """
-    Import the specified module. If the module is not installed,
-    raise an ImportError with a helpful message.
-
-    Parameters
-    ----------
-    module : str
-        The name of the module to import
-    mitigation : Optional[str]
-        The package(s) to install to mitigate the error.
-        If not provided then the module name will be used.
-    """
-    try:
-        return importlib.import_module(module)
-    except ImportError:
-        raise ImportError(f"Please install {mitigation or module}")
